simplify the merge story with another PR
